requirements.txt
pyyaml
aiohttp
aiofiles
bidict
python-novaclient==10.0.0
python-glanceclient==2.9.1
keystoneauth1==3.4.0
python-keystoneclient==3.10.0
